+++ frontend/ai.client/src/app/agents/agent-form/agent-form.page.html
@@ -244,17 +244,12 @@
       </section>
     }
 
-    <!-- Knowledge base (welded, read-only) -->
-    @if (kbBinding()) {
-      <section class="rounded-2xl border border-gray-200/80 bg-white p-6 shadow-xs dark:border-gray-700/60 dark:bg-gray-800/80">
-        <div class="flex items-center gap-2">
-          <ng-icon name="heroBookOpen" class="size-5 text-gray-400" aria-hidden="true" />
-          <h2 class="text-sm/6 font-semibold text-gray-900 dark:text-white">Knowledge base</h2>
-        </div>
-        <p class="mt-1 text-xs/5 text-gray-500 dark:text-gray-400">
-          This agent's document knowledge base is managed automatically and is edited from the classic assistant view.
-        </p>
-      </section>
-    }
+    <!-- Knowledge base (shared component — also used by the assistant editor) -->
+    <app-knowledge-base-section
+      [entityId]="agentId()"
+      [userPermission]="userPermission()"
+      [permissionResolved]="permissionResolved()"
+      [createDraft]="createDraftAgent"
+    />
   </form>
 </div>
